+++ catalog/controller/payment/nuvei.php
@@ -91,6 +91,8 @@ class Nuvei extends \Opencart\System\Engine\Controller
             $useDCC = 'false';
         }
         
+        $locale = substr($this->get_locale(), 0, 2);
+        
         $data['nuvei_sdk_params'] = [
             'renderTo'                  => '#nuvei_checkout',
             'strict'                    => false,
@@ -113,11 +115,16 @@ class Nuvei extends \Opencart\System\Engine\Controller
             'fullName'                  => trim($order_data['billingAddress']['firstName'] 
                 . ' ' . $order_data['billingAddress']['lastName']),
             'payButton'                 => $this->plugin_settings[NUVEI_SETTINGS_PREFIX . 'pay_btn_text'],
-            'locale'                    => substr($this->get_locale(), 0, 2),
+            'locale'                    => $locale,
             'autoOpenPM'                => (bool) $this->plugin_settings[NUVEI_SETTINGS_PREFIX . 'auto_expand_pms'],
             'logLevel'                  => $this->plugin_settings[NUVEI_SETTINGS_PREFIX . 'sdk_log_level'],
             'i18n'                      => json_decode($this->plugin_settings[NUVEI_SETTINGS_PREFIX . 'sdk_transl'], true),
             'theme'                     => $this->plugin_settings[NUVEI_SETTINGS_PREFIX . 'sdk_theme'],
+            'apmConfig'                 => [
+                'googlePay' => [
+                    'locale' => $locale
+                ]
+            ],
 //            'apmWindowType'             => $this->plugin_settings[NUVEI_SETTINGS_PREFIX . 'apm_window_type'],
         ];